assignment/requirements.txt
streamlit>=1.36.0
google-generativeai>=0.7.0
pytz>=2023.3

